fix: Update import paths after kebab-case renaming
