+++ src/modules/availability/routes/working-hours-routes.ts
@@ -1,6 +1,7 @@
 import { FastifyInstance, FastifyReply, FastifyRequest } from "fastify";
 
 import { UpdateWorkingHoursBody } from "../types";
+
 import { AVAILABILITY_CONSTANTS } from "../constants/availability";
 
 import { WorkingHoursController } from "../controllers/working-hours-controller";
